General increase of idempotency + handling of mishaps when setting wrong sha256sum
This is currently the state of the role that I have been using myself in
a forked repository. Some notable points:

- Added 4 variables to defaults/main.yml to simplify line lengths /
undertanding in playbooks
- Added an install.yml in tasks/ as a handler for install related
tasks (where upgrade.yml still handles backup-related)
- Moved the group check after the user check, because if the user is
created it is created by default with its group, so the check is
essentially causing the group/user to diverge because it creates the
group first, then the user
- Instead of using a static name for the new tarball, we take the
short sha256 and append that to the name - this prevents issues when you
don't set the sha256sum properly during an update
- Instead of blindly clobbering the backup, check to see if we've
backed up in the same day and don't move the current bundle into a
backup if we have.
    - This replaces the "Delete the current..." task