Add ensure acl-tools is present on Ubuntu 16
Ubuntu 16 is missing sefacl by default and needs the `acl` package in
order to cooperate with become_user and Ansible 2.1, otherwise it fails
with the following:

    TASK [../../RocketChat.Server : Fetch the Rocket.Chat binary
tarball] **********
    fatal: [ubuntu16]: FAILED! => {"failed": true, "msg": "Failed to set
permissions on the temporary files Ansible needs to create when becoming
an unprivileged user. For information on working around this, see
https://docs.ansible.com/ansible/become.html#becoming-an-unprivileged-user"}

I followed the second recommendation as the best workaround and tested
acl-tools as a working solution.