Remove forever daemon in lieu of plain Upstart on Ubuntu 14
The previous method of using the forever monitor daemon with upstart
failed at being able to write to the pid-file upon reboot. This showed
up in Vagrant testing and I believe it has something to do with the fact
that we're creating the pidfile manually within the play.

I have prior experience with NodeJS+Upstart and it turns out that
forever isn't actually necessary at all. With `console log` and
`respawn` directives, and an unforked process, Upstart properly tracks
the PID (without a pidfile), respawns the process should it die, and
logs by default to `/var/log/upstart/rocketchat.log`

This changeset removes the following tasks and has the added benefit of
simplifying the play overall:

- Ensure Forever (NodeJS) is installed [Ubuntu 14]
- Check to see if the Rocket.Chat log file is present [Ubuntu 14]
- Ensure the Rocket.Chat log file is present [Ubuntu 14]
- Ensure the Rocket.Chat pid file is present [Ubuntu 14]

The former log-file path has been replaced with a symlink pointing to
the new logfile path.

This fixes #11